в бд поставил на баланс тип numeric

+++ src/main/resources/liquibase/scripts/createTables.sql
@@ -5,8 +5,8 @@ create table accounts
 (
     id         BigSerial      primary key,
     name       varchar     not null,
-    balance double     not null,
-    pin_code varchar not null
+    balance    numeric     not null,
+    pin_code   varchar not null
 );
 INSERT INTO accounts (id, name, balance, pin_code) VALUES
     (100, 'first', 100.99, 1111),
